split up read/write in file adapter tests
the length of the original file.spec.js was getting a bit out of hand
and this will help with maintaining these tests a bit easier
